pyproject.toml
[tool.poetry]
name = "brainiac"
version = "0.1.0"
description = "Local-first, uncensored AI chatbot optimized for 8GB VRAM"
authors = ["brAIniac Team"]
readme = "README.md"
license = "CC0-1.0"
packages = [
    { include = "core" },
    { include = "servers" },
]

[tool.poetry.dependencies]
python = "~3.12"
ollama = "^0.4.4"
fastmcp = "^2.2.0"
python-dotenv = "^1.0.1"
rich = "^13.9.4"
ddgs = "^9.10.0"
gradio = "^6.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
pytest-mock = "^3.15.1"
pytest-cov = "^7.0.0"
black = "^26.1.0"
mypy = "^1.19.1"
ruff = "^0.15.2"
isort = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["core", "servers"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
import_heading_stdlib = "Standard Library"
import_heading_thirdparty = "Third-Party Libraries"
import_heading_firstparty = "Local Modules"
length_sort = true

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--strict-markers",
    "--cov=core",
    "--cov=servers",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests that require a live Ollama instance",
]

[tool.coverage.run]
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/conftest.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "\\.\\.\\.",
]
